updating structure
