+++ Makefile
@@ -8,3 +8,6 @@ fmt:
 	cargo fmt --all --check
 
 build: clippy test fmt
+
+run_debug:
+	POWERTOP_CONFIG=`pwd`/.config POWERTOP_DATA=`pwd`/.data POWERTOP_LOG_LEVEL=debug cargo run
