+++ src/configuration/configuration.rs
@@ -1,0 +1,63 @@
+use std::path::PathBuf;
+
+use color_eyre::eyre::Result;
+use directories::ProjectDirs;
+use lazy_static::lazy_static;
+use tracing::error;
+use tracing_error::ErrorLayer;
+use tracing_subscriber::{self, prelude::__tracing_subscriber_SubscriberExt, util::SubscriberInitExt, Layer};
+
+pub const VERSION_MESSAGE: &str =
+  concat!(env!("CARGO_PKG_VERSION"), "-", env!("VERGEN_GIT_DESCRIBE"), " (", env!("VERGEN_BUILD_DATE"), ")");
+
+lazy_static! {
+  pub static ref PROJECT_NAME: String = env!("CARGO_CRATE_NAME").to_uppercase().to_string();
+  pub static ref DATA_FOLDER: Option<PathBuf> =
+    std::env::var(format!("{}_DATA", PROJECT_NAME.clone())).ok().map(PathBuf::from);
+  pub static ref CONFIG_FOLDER: Option<PathBuf> =
+    std::env::var(format!("{}_CONFIG", PROJECT_NAME.clone())).ok().map(PathBuf::from);
+}
+
+fn project_directory() -> Option<ProjectDirs> {
+  ProjectDirs::from("com", "jteer", env!("CARGO_PKG_NAME"))
+}
+
+pub fn get_data_dir() -> PathBuf {
+  let directory = if let Some(s) = DATA_FOLDER.clone() {
+    s
+  } else if let Some(proj_dirs) = project_directory() {
+    proj_dirs.data_local_dir().to_path_buf()
+  } else {
+    PathBuf::from(".").join(".data")
+  };
+  directory
+}
+
+pub fn get_config_dir() -> PathBuf {
+  let directory = if let Some(s) = CONFIG_FOLDER.clone() {
+    s
+  } else if let Some(proj_dirs) = project_directory() {
+    proj_dirs.config_local_dir().to_path_buf()
+  } else {
+    PathBuf::from(".").join(".config")
+  };
+  directory
+}
+
+pub fn version() -> String {
+  let author = clap::crate_authors!();
+
+  // let current_exe_path = PathBuf::from(clap::crate_name!()).display().to_string();
+  let config_dir_path = get_config_dir().display().to_string();
+  let data_dir_path = get_data_dir().display().to_string();
+
+  format!(
+    "\
+  {VERSION_MESSAGE}
+  
+  Authors: {author}
+  
+  Config directory: {config_dir_path}
+  Data directory: {data_dir_path}"
+  )
+}
