+++ src/main.rs
@@ -11,17 +11,20 @@ pub mod data_services;
 pub mod mode;
 pub mod tui;
 pub mod utils;
+pub mod logging;
+pub mod configuration;
 
 use std::sync::{Arc, Mutex};
 
 use clap::Parser;
 use cli::Cli;
 use color_eyre::eyre::Result;
+use logging::logging::initialize_logging;
 use sysinfo::System;
 
 use crate::{
   app::App,
-  utils::{initialize_logging, initialize_panic_handler, version},
+  utils::initialize_panic_handler,
 };
 
 async fn tokio_main() -> Result<()> {
